clone bug fix: #2377
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@6535 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/java/org.wcs.smart.report/src/org/wcs/smart/report/ReportTemplateCloner.java
@@ -163,15 +163,17 @@ public class ReportTemplateCloner implements
 			File src = new File(new File(engine.getTemplateCa().getFileDataStoreLocation(), Report.REPORT_DIR), r.getFilename());
 			File dest = new File(new File(engine.getNewCa().getFileDataStoreLocation(), Report.REPORT_DIR), clone.getFilename());
 			
-			FileUtils.copyFile(src, dest);
-	
-			boolean save = true;
-			try{
-				updateReportFile(clone, dest,  engine);	
-			}catch (Exception ex){
-				save = false;
-				ReportPlugIn.log(ex.getMessage(), ex);
-				dest.delete();
+			boolean save = false;
+			if (src.exists()) {
+				FileUtils.copyFile(src, dest);
+				save = true;
+				try{
+					updateReportFile(clone, dest,  engine);	
+				}catch (Exception ex){
+					save = false;
+					ReportPlugIn.log(ex.getMessage(), ex);
+					dest.delete();
+				}
 			}
 			
 			if (save){
